Use Swift's native error handling instead of catching errors

+++ Tella/Data/Networking/Upload/BaseUploadOperation.swift
@@ -117,12 +117,8 @@ class BaseUploadOperation : Operation {
         let report = Report(id: self.report?.id,
                             status: reportStatus,
                             apiID: apiID)
-        do {
-            try mainAppModel.vaultManager.tellaData?.updateReport(report: report)
-            
-        } catch {
-            
-        }
+        
+        mainAppModel.vaultManager.tellaData?.updateReport(report: report)
     }
     
     func updateReportFile(fileStatus:FileStatus, id:Int?, bytesSent:Int? = nil, current:Int? = nil ) -> Int {
@@ -146,13 +142,9 @@ class BaseUploadOperation : Operation {
         let file = self.reportVaultFiles?.first(where: {$0.instanceId == id})
         let totalBytesSent = (file?.current ?? 0)  + (file?.bytesSent ?? 0)
         
-        do {
-            try mainAppModel.vaultManager.tellaData?.updateReportFile(reportFile: ReportFile(id: id,
-                                                                                             status: fileStatus,
-                                                                                             bytesSent: totalBytesSent))
-        } catch {
-            
-        }
+        mainAppModel.vaultManager.tellaData?.updateReportFile(reportFile: ReportFile(id: id,
+                                                                                     status: fileStatus,
+                                                                                     bytesSent: totalBytesSent))
         return totalBytesSent
     }
     
@@ -180,10 +172,13 @@ class BaseUploadOperation : Operation {
     }
     
     func deleteCurrentAutoReport() {
-        mainAppModel.vaultManager.tellaData?.deleteReport(reportId: self.report?.id)
-        guard let reportVaultFiles = self.reportVaultFiles else {return}
-        let reportVaultFilesIds = reportVaultFiles.compactMap{ $0.id}
-        mainAppModel.vaultFilesManager?.deleteVaultFile(fileIds: reportVaultFilesIds) 
+        let deleteReportResult = mainAppModel.vaultManager.tellaData?.deleteReport(reportId: self.report?.id)
+        
+        if case .success = deleteReportResult {
+            guard let reportVaultFiles = self.reportVaultFiles else {return}
+            let reportVaultFilesIds = reportVaultFiles.compactMap{ $0.id}
+            mainAppModel.vaultFilesManager?.deleteVaultFile(fileIds: reportVaultFilesIds)
+        }
     }
     
     func sendReport() {
@@ -216,7 +211,7 @@ class BaseUploadOperation : Operation {
             
             guard let apiID = self.report?.apiID, let accessToken = report?.server?.accessToken, let serverUrl = report?.server?.url else { return }
             
-            if let filesToUpload = reportVaultFiles?.filter{$0.status != .submitted} {
+            if let filesToUpload = reportVaultFiles?.filter({$0.status != .submitted}) {
                 if filesToUpload.isEmpty {
                     self.checkAllFilesAreUploaded()
                 } else {
